patch trajectoire

+++ trajectoire.java
@@ -27,8 +27,8 @@ public class trajectoire {
 
             // taille de tableaux = distance horizontale max
             if (i == 0) {
-                for (Integer absciss = 0; absciss < maxX; absciss++) {
-                    paramTraj[0].add(absciss, absciss);
+                for (double absciss = 0; absciss < maxX; absciss++) {
+                    paramTraj[0].add((int) absciss, (int) absciss);
                     y = (int) (-0.5 * gravity / (speedInit * speedInit) * absciss * absciss
                             * (1 + Math.pow(Math.tan(angleInit), 2)) + absciss * Math.tan(angleInit));
                     paramTraj[1].add(y);
